pyproject.toml
[tool.poetry]
name = "abuse_whois"
version = "0.4.1"
description = "Find where to report a domain for abuse"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"
homepage = "https://github.com/ninoseki/abuse_whois"
repository = "https://github.com/ninoseki/abuse_whois"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiometer = "^0.4.0"
asyncache = "^0.3.1"
asyncer = "^0.0.2"
cachetools = "^5.2.1"
email-validator = "^1.3.1"
fastapi = "^0.89.1"
gunicorn = { version = "^20.1.0", optional = true }
lark-parser = "^0.12.0"
loguru = "^0.6.0"
pydantic = "^1.10.4"
pyhumps = "^3.8.0"
PyYAML = "^6.0"
regex = "^2022.10.31"
tldextract = "^3.4.0"
typer = "^0.7.0"
uvicorn = { extras = ["standard"], version = "^0.20.0", optional = true }
whois-parser = "^0.1.4"

[tool.poetry.dev-dependencies]
autoflake = "^2.0"
autopep8 = "^2.0"
black = "^22.12"
coveralls = "^3.3"
isort = "^5.11"
mypy = "^0.991"
pre-commit = "^2.21"
pytest = "^7.2"
pytest-asyncio = "^0.20"
pytest-cov = "^4.0"
pytest-mock = "^3.10"
pytest-randomly = "^3.12"
pytest-sugar = "^0.9"
pyupgrade = "^3.3"

[tool.poetry.extras]
api = ["uvicorn", "gunicorn"]

[tool.poetry.scripts]
abuse_whois = "abuse_whois.cli:app"

[build-system]
requires = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
